+++ spec/view_helper/title_spec.rb
@@ -1,101 +1,112 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe MetaTags::ViewHelper do
   subject { ActionView::Base.new }
 
-  context 'displaying title' do
-    it 'should not display title if blank' do
+  describe 'displaying title' do
+    it 'does not display title if blank' do
       expect(subject.display_meta_tags).to eq('')
       subject.page_title('')
       expect(subject.display_meta_tags).to eq('')
     end
 
-    it 'should use website name if title is empty' do
+    it 'uses website name if title is empty' do
       subject.display_meta_tags(site: 'someSite').tap do |meta|
         expect(meta).to eq('<title>someSite</title>')
       end
     end
 
-    it 'should display title when "title" used' do
+    it 'displays title when "title" used' do
       subject.page_title('someTitle')
       subject.display_meta_tags(site: 'someSite').tap do |meta|
         expect(meta).to eq('<title>someSite | someTitle</title>')
       end
     end
 
-    it 'should display title only when "site" is empty' do
+    it 'displays title only when "site" is empty' do
       subject.page_title('someTitle')
       expect(subject.display_meta_tags).to eq('<title>someTitle</title>')
     end
 
-    it 'should display title when "set_meta_tags" is used' do
+    it 'displays title when "set_meta_tags" is used' do
       subject.set_meta_tags(title: 'someTitle')
       subject.display_meta_tags(site: 'someSite').tap do |meta|
         expect(meta).to eq('<title>someSite | someTitle</title>')
       end
     end
 
-    it 'should escape the title when "set_meta_tags" is used' do
+    it 'escapes the title when "set_meta_tags" is used' do
       subject.set_meta_tags(title: 'someTitle & somethingElse')
       subject.display_meta_tags(site: 'someSite').tap do |meta|
         expect(meta).to eq('<title>someSite | someTitle &amp; somethingElse</title>')
       end
     end
 
-    it 'should escape a very long title when "set_meta_tags" is used' do
+    it 'escapes a very long title when "set_meta_tags" is used' do
       subject.set_meta_tags(title: 'Kombucha kale chips forage try-hard & green juice. IPhone marfa PBR&B venmo listicle, irony kitsch thundercats.')
       subject.display_meta_tags(site: 'someSite').tap do |meta|
-        expect(meta).to eq('<title>someSite | Kombucha kale chips forage try-hard &amp; green juice.</title>')
+        expect(meta).to eq('<title>someSite | Kombucha kale chips forage try-hard &amp; green juice. IPhone</title>')
       end
     end
 
-    it 'should strip tags in the title' do
+    it 'strips tags in the title' do
       subject.set_meta_tags(title: '<b>hackxor</b>')
       subject.display_meta_tags(site: 'someSite').tap do |meta|
         expect(meta).to eq('<title>someSite | hackxor</title>')
       end
     end
 
-    it 'should strip tags from very long titles' do
+    it 'strips tags from very long titles' do
       subject.set_meta_tags(title: 'Kombucha <b>kale</b> chips forage try-hard & green juice. IPhone marfa PBR&B venmo listicle, irony kitsch thundercats.')
       subject.display_meta_tags(site: 'someSite').tap do |meta|
-        expect(meta).to eq('<title>someSite | Kombucha kale chips forage try-hard &amp; green juice.</title>')
+        expect(meta).to eq('<title>someSite | Kombucha kale chips forage try-hard &amp; green juice. IPhone</title>')
       end
     end
 
-    it 'should display custom title if given' do
+    it 'displays custom title if given' do
       subject.page_title('someTitle')
       subject.display_meta_tags(site: 'someSite', title: 'defaultTitle').tap do |meta|
         expect(meta).to eq('<title>someSite | someTitle</title>')
       end
     end
 
-    it 'should use website before page by default' do
+    it 'uses website before page by default' do
       subject.display_meta_tags(site: 'someSite', title: 'someTitle').tap do |meta|
         expect(meta).to eq('<title>someSite | someTitle</title>')
       end
     end
 
-    it 'should only use markup in titles in the view' do
+    it 'only uses markup in titles in the view' do
       expect(subject.page_title('<b>someTitle</b>')).to eq('<b>someTitle</b>')
       subject.display_meta_tags(site: 'someSite').tap do |meta|
         expect(meta).to eq('<title>someSite | someTitle</title>')
       end
     end
 
-    it 'should use page before website if :reverse' do
+    it 'uses page before website if :reverse' do
       subject.display_meta_tags(site: 'someSite', title: 'someTitle', reverse: true).tap do |meta|
         expect(meta).to eq('<title>someTitle | someSite</title>')
       end
     end
 
-    it 'should be lowercase if :lowercase' do
+    it 'is lowercase if :lowercase' do
       subject.display_meta_tags(site: 'someSite', title: 'someTitle', lowercase: true).tap do |meta|
         expect(meta).to eq('<title>someSite | sometitle</title>')
       end
     end
 
-    it 'should use custom separator if :separator' do
+    it 'does not change original title string' do
+      title = "TITLE"
+      subject.display_meta_tags(title: title, lowercase: true).tap do |meta|
+        expect(meta).to eq('<title>title</title>')
+      end
+      expect(title).to eq('TITLE')
+    end
+
+    # rubocop:disable Rails/OutputSafety
+    it 'uses custom separator when :separator specified' do
       subject.page_title('someTitle')
       subject.display_meta_tags(site: 'someSite', separator: '-').tap do |meta|
         expect(meta).to eq('<title>someSite - someTitle</title>')
@@ -106,33 +117,37 @@ describe MetaTags::ViewHelper do
       subject.display_meta_tags(site: 'someSite', separator: '&amp;').tap do |meta|
         expect(meta).to eq('<title>someSite &amp;amp; someTitle</title>')
       end
+      subject.display_meta_tags(site: 'someSite', separator: '&').tap do |meta|
+        expect(meta).to eq('<title>someSite &amp; someTitle</title>')
+      end
       subject.display_meta_tags(site: 'someSite', separator: '&amp;'.html_safe).tap do |meta|
         expect(meta).to eq('<title>someSite &amp; someTitle</title>')
       end
-      subject.display_meta_tags(site: 'someSite: ', separator: false).tap do |meta|
-        expect(meta).to eq('<title>someSite: someTitle</title>')
+      subject.display_meta_tags(site: 'someSite:', separator: false).tap do |meta|
+        expect(meta).to eq('<title>someSite:someTitle</title>')
       end
     end
+    # rubocop:enable Rails/OutputSafety
 
-    it 'should use custom prefix and suffix if available' do
+    it 'uses custom prefix and suffix if available' do
       subject.display_meta_tags(site: 'someSite', title: 'someTitle', prefix: ' -', suffix: '- ').tap do |meta|
         expect(meta).to eq('<title>someSite -|- someTitle</title>')
       end
     end
 
-    it 'should collapse prefix if false' do
+    it 'collapses prefix if false' do
       subject.display_meta_tags(site: 'someSite', title: 'someTitle', prefix: false).tap do |meta|
         expect(meta).to eq('<title>someSite| someTitle</title>')
       end
     end
 
-    it 'should collapse suffix if false' do
+    it 'collapses suffix if false' do
       subject.display_meta_tags(site: 'someSite', title: 'someTitle', suffix: false).tap do |meta|
         expect(meta).to eq('<title>someSite |someTitle</title>')
       end
     end
 
-    it 'should use all custom options if available' do
+    it 'uses all custom options if available' do
       subject.display_meta_tags(
         site:      'someSite',
         title:     'someTitle',
@@ -146,19 +161,37 @@ describe MetaTags::ViewHelper do
       end
     end
 
-    it 'should allow Arrays in title' do
+    it 'allows Arrays in title' do
       subject.display_meta_tags(site: 'someSite', title: ['someTitle', 'anotherTitle']).tap do |meta|
         expect(meta).to eq('<title>someSite | someTitle | anotherTitle</title>')
       end
     end
 
-    it 'should allow Arrays in title with :lowercase' do
+    it 'allows Arrays in title with :lowercase' do
       subject.display_meta_tags(site: 'someSite', title: ['someTitle', 'anotherTitle'], lowercase: true).tap do |meta|
         expect(meta).to eq('<title>someSite | sometitle | anothertitle</title>')
       end
     end
 
-    it 'should build title in reverse order if :reverse' do
+    it 'treats nil as an empty string' do
+      subject.display_meta_tags(title: nil).tap do |meta|
+        expect(meta).not_to have_tag('title')
+      end
+    end
+
+    it 'allows objects that respond to #to_str' do
+      title = double(to_str: 'someTitle')
+      subject.display_meta_tags(site: 'someSite', title: title).tap do |meta|
+        expect(meta).to eq('<title>someSite | someTitle</title>')
+      end
+    end
+
+    it 'fails when title is not a String-like object' do
+      expect { subject.display_meta_tags(site: 'someSite', title: 5) }.to \
+        raise_error ArgumentError, 'Expected a string or an object that implements #to_str'
+    end
+
+    it 'builds title in reverse order if :reverse' do
       subject.display_meta_tags(
         site:      'someSite',
         title:     ['someTitle', 'anotherTitle'],
@@ -170,10 +203,21 @@ describe MetaTags::ViewHelper do
         expect(meta).to eq('<title>anotherTitle -:+ someTitle -:+ someSite</title>')
       end
     end
+
+    it 'minifies the output when asked to' do
+      subject.display_meta_tags(title: 'hello', description: 'world').tap do |meta|
+        expect(meta).to eq("<title>hello</title>\n<meta name=\"description\" content=\"world\">")
+      end
+
+      MetaTags.config.minify_output = true
+      subject.display_meta_tags(title: 'hello', description: 'world').tap do |meta|
+        expect(meta).to eq("<title>hello</title><meta name=\"description\" content=\"world\">")
+      end
+    end
   end
 
-  context '.display_title' do
-    it 'should display custom title if given' do
+  describe '.display_title' do
+    it 'displays custom title if given' do
       subject.page_title('someTitle')
       subject.display_title(site: 'someSite', title: 'defaultTitle').tap do |meta|
         expect(meta).to eq('someSite | someTitle')